repair: extract repair_writer
