+++ configure.py
@@ -858,6 +858,7 @@ scylla_core = (['database.cc',
                 'repair/sink_source_for_repair.cc',
                 'repair/row.cc',
                 'repair/reader.cc',
+                'repair/writer.cc',
                 'repair/repair.cc',
                 'repair/row_level.cc',
                 'exceptions/exceptions.cc',
